+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - [\#376](https://github.com/cosmos/evm/pull/376) Fix precompile initialization for local node development script
 - [\#384](https://github.com/cosmos/evm/pull/384) Fix debug_traceTransaction RPC failing with block height mismatch errors
 - [\#441](https://github.com/cosmos/evm/pull/441) Align precompiles map with available static check to Prague.
+- [\#454](https://github.com/cosmos/evm/pull/454) Align multi decode functions instead of string contains check in HexAddressFromBech32String.
 
 ### IMPROVEMENTS